Factor out common week 2 code to simplify code students need to edit
